address Jiayin's comment to move trigger with unknown user type event handler from infoMessageObserver to errorMessageObserver
