+++ src/observers/errorMessageObserver.ts
@@ -1,7 +1,7 @@
 import vscode from 'vscode';
 import { EventType } from '../common/eventType';
 import { BaseEvent, UserSignInFailed, UserSignInCompleted, UserSignOutCompleted, UserSignOutFailed, BuildCompleted, BuildFailed } from '../common/loggingEvents';
-import { MessageAction } from '../shared';
+import { MessageAction, EXTENSION_NAME, USER_TYPE, UserType } from '../shared';
 import { ErrorCode } from '../error/errorCode';
 import { DocsError } from '../error/docsError';
 import { DocfxExecutionResult } from '../build/buildResult';
@@ -29,15 +29,20 @@ export class ErrorMessageObserver {
             case EventType.PublicUserSignIn:
                 this.handlePublicUserSignIn();
                 break;
+            case EventType.TriggerCommandWithUnkownUserType:
+                this.handleCommandWithUnkownUserTypeTriggered();
+                break;
         }
     }
 
-    private async showErrorMessage(message: string, action?: MessageAction) {
-        let infoMsg = `[Docs Validation] ${message}`;
-        if (action && action.description) {
-            infoMsg += `\n${action.description}`;
-        }
-        let input = <MessageAction>(await vscode.window.showErrorMessage(infoMsg, action));
+    private async showErrorMessage(message: string, ...actions: MessageAction[]) {
+        let errorMsg = `[Docs Validation] ${message}`;
+        actions.forEach((action) => {
+            if (action && action.description) {
+                errorMsg += ` ${action.description}`;
+            }
+        });
+        let input = <MessageAction>(await vscode.window.showErrorMessage(errorMsg, ...actions));
         if (input) {
             if (input.command) {
                 vscode.commands.executeCommand(input.command, undefined);
@@ -60,5 +65,29 @@ export class ErrorMessageObserver {
 
     private handlePublicUserSignIn() {
         this.showErrorMessage(`Sign in is only available for Microsoft employees.`);
+    }
+
+    private handleCommandWithUnkownUserTypeTriggered() {
+        this.showErrorMessage(
+            `The command you triggered needs user type information. Please choose either Microsoft employee or Public contributor. ` +
+            `You can change your selection later if needed in the extension settings (Docs validation -> User type).`,
+            new MessageAction(
+                "Microsoft employee",
+                undefined,
+                undefined,
+                () => {
+                    const extensionConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(EXTENSION_NAME, undefined);
+                    extensionConfig.update(USER_TYPE, UserType.MicrosoftEmployee, true);
+                }
+            ),
+            new MessageAction(
+                "Public contributor",
+                undefined,
+                undefined,
+                () => {
+                    const extensionConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(EXTENSION_NAME, undefined);
+                    extensionConfig.update(USER_TYPE, UserType.PublicContributor, true);
+                }
+            ));
     }
 }
